+++ package.json
@@ -103,13 +103,6 @@
     "url": "https://github.com/TypeStrong/ts-node/issues"
   },
   "homepage": "https://typestrong.org/ts-node",
-  "ava": {
-    "files": [
-      "dist/test/*.spec.js"
-    ],
-    "failWithoutAssertions": false,
-    "timeout": "300s"
-  },
   "devDependencies": {
     "@microsoft/api-extractor": "^7.15.2",
     "@swc/core": ">=1.2.50",
@@ -118,6 +111,7 @@
     "@types/diff": "^4.0.2",
     "@types/lodash": "^4.14.151",
     "@types/node": "13.13.5",
+    "@types/proper-lockfile": "^4.1.2",
     "@types/proxyquire": "^1.3.28",
     "@types/react": "^16.0.2",
     "@types/rimraf": "^3.0.0",
@@ -132,6 +126,7 @@
     "ntypescript": "^1.201507091536.1",
     "nyc": "^15.0.1",
     "prettier": "^2.2.1",
+    "proper-lockfile": "^4.1.2",
     "proxyquire": "^2.0.0",
     "react": "^16.14.0",
     "rimraf": "^3.0.0",
